allow extra app dependencies for yaws.app to be configured
The default yaws.app file specifies only kernel and stdlib as application
dependencies, but depending on how it's used, Yaws can also depend on the
crypto and compiler applications. If these are used but are missing from
the yaws.app file, building an Erlang/OTP release will fail.

Add --enable-crypto and --enable-compiler options to the configure script
to allow the user to add either or both of these applications to the
yaws.app file. Specifying the --enable-crypto option also causes the Yaws
startup code to start the crypto application; likewise for
--enable-compiler and the compiler application.